+++ src/main/java/grisu/gricli/command/InteractiveLoginCommand.java
@@ -49,11 +49,6 @@ GricliCommand {
 
 		new SetCommand("application", Constants.GENERIC_APPLICATION_NAME)
 		.execute(env);
-		String lastVo = ClientPropertiesManager.getLastUsedFqan();
-		if (StringUtils.isNotBlank(lastVo)) {
-			env.printMessage("Setting last used group: " + lastVo);
-			new SetCommand("group", lastVo).execute(env);
-		}
 
 		String value = System
 				.getenv(LocalLoginCommand.GRICLI_LOGIN_SCRIPT_ENV_NAME);