+++ src/main/java/grisu/gricli/command/LocalLoginCommand.java
@@ -50,11 +50,6 @@ GricliCommand {
 
 			new SetCommand("application", Constants.GENERIC_APPLICATION_NAME)
 			.execute(env);
-			String lastVo = ClientPropertiesManager.getLastUsedFqan();
-			if (StringUtils.isNotBlank(lastVo)) {
-				env.printMessage("Setting last used group: " + lastVo);
-				new SetCommand("group", lastVo).execute(env);
-			}
 
 			String value = System.getenv(GRICLI_LOGIN_SCRIPT_ENV_NAME);
 			if (StringUtils.isNotBlank(value)) {