+++ src/test/java/grisu/gricli/TestCommands.java
@@ -1,14 +1,13 @@
 package grisu.gricli;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+import static org.junit.Assert.*;
 import grisu.gricli.command.AddCommand;
 import grisu.gricli.command.AttachCommand;
 import grisu.gricli.command.ChdirCommand;
 import grisu.gricli.command.GricliCommandFactory;
 import grisu.gricli.command.RunCommand;
 import grisu.gricli.command.SetCommand;
+import grisu.gricli.command.SubmitCommand;
 import grisu.gricli.environment.GricliEnvironment;
 
 import java.io.File;
@@ -194,10 +193,22 @@ public class TestCommands {
 
 		assertEquals(env.jobname.get(),"hello");
 	}
+	
+	@Test
+	public void testRunWithComments() throws Exception {
+		List<String> script = new LinkedList<String>();
+		script.add("# this is comment");
+		File f = folder.newFile("testRun2.script");
+		String scriptName = f.getCanonicalPath();
+		FileUtils.writeLines(f,script);
+		
+		RunCommand c = new RunCommand(scriptName);
+		c.execute(env);
+	}
 
 	@Test
 	public void testRunWithNoEndOfLine() throws Exception {
-		File f = folder.newFile("testRun2.script");
+		File f = folder.newFile("testRun3.script");
 		String scriptName = f.getCanonicalPath();
 		FileUtils.writeByteArrayToFile(f, "set jobname hello".getBytes());
 
@@ -244,6 +255,54 @@ public class TestCommands {
 	@Test(expected=GricliSetValueException.class)
 	public void testSetJobNameWithSpaces() throws Exception {
 		env.jobname.set("job name with spaces");
+	}
+	
+	// testing submit commands
+	
+	@Test
+	public void testSimpleSubmitCmd() throws Exception {
+		SubmitCommand submit = new SubmitCommand("java","-version");
+		assertEquals("java -version", submit.getCommandline());
+	}
+	
+	@Test
+	public void testSubmitCmdWithTilda() throws Exception {
+		SubmitCommand submit = new SubmitCommand("java","-version", "&");
+		assertEquals("java -version", submit.getCommandline());
+	}
+	
+	@Test
+	public void testSubmitWithSpaces() throws Exception {
+		SubmitCommand submit = new SubmitCommand("cat","file with spaces");
+		assertEquals("cat \"file with spaces\"", submit.getCommandline());
+	}
+	
+	@Test
+	public void testSubmitWithQuotes() throws Exception {
+		SubmitCommand submit = new SubmitCommand("crazyquotes","\"a\"");
+		assertEquals("crazyquotes \"\\\"a\\\"\"", submit.getCommandline());
+	}
+	
+	// test set and unset commands
+	
+	@Test
+	public void testUnsetRightVar() throws Exception{
+		SetCommand unset = new SetCommand("queue");
+		unset.execute(env);
+		assertNull(env.queue.get());
+	}
+	
+	@Test(expected=GricliSetValueException.class)
+	public void testUnsetWrongVar() throws Exception {
+		SetCommand unset = new SetCommand("cpus");
+		unset.execute(env);
+	}
+	
+	@Test
+	public void testUnsetHostCount() throws Exception{
+		SetCommand unset = new SetCommand("hostCount");
+		unset.execute(env);
+		assertNull(env.queue.get());
 	}
 
 }