+++ src/test/java/grisu/gricli/parser/TestParsing.java
@@ -93,5 +93,50 @@ public class TestParsing {
 		assertArrayEquals(tokenize("attach ~/test"),new String[] {"attach", "~/test"});
 	}
 	
+	// test line numbers
+	@Test
+	public void testFirstLine() throws Exception{
+		tokenizer = str2Tokenizer("aaa bbb");
+		assertEquals(tokenizer.getLineNumber(),0);
+		tokenizer.nextCommand();
+		assertEquals(tokenizer.getLineNumber(),0);
+	}
+	
+	@Test
+	public void testSecondLine() throws Exception{
+		tokenizer = str2Tokenizer("aaa bbb\n");
+		tokenizer.nextCommand();
+		assertEquals(tokenizer.getLineNumber(),1);
+	}
+	
+	@Test
+	public void testThirdLine() throws Exception {
+		tokenizer = str2Tokenizer("aaa bbb\naaa bbb\n");
+		tokenizer.nextCommand();
+		tokenizer.nextCommand();
+		assertEquals(tokenizer.getLineNumber(),2);
+	}
+	
+	// test escape
+	@Test
+	public void testNullEscape(){
+		assertNull(GricliTokenizer.escape(null));
+	}
+	
+	@Test
+	public void testTrivialEscape(){
+		assertEquals("a",GricliTokenizer.escape("a"));
+	}
+	
+	@Test
+	public void testSpaceEscape(){
+		assertEquals("\"a b\"",GricliTokenizer.escape("a b"));
+	}
+	
+	@Test
+	public void testEmptyLineEscape(){
+		assertEquals("\"\"",GricliTokenizer.escape(""));
+	}
+	
 
 }