+++ CHANGELOG.md
@@ -4,8 +4,14 @@ Changelog
 0.5
 ---
 
+* script comments #
+* glob support for print package
+* new commands clean/kill jobs
+* persistent user session; globals except attached files and env vars saved
+* last used group remembered
 * tab-completion for remote file systems
 * tab-completion now also works on "submit" command for executables and input files
+* submit command no longer needs quotes
 * BeSTGRID-DEV renamed to **dev** case insensitive
 * "print queues" now displays only queues that are available for the currently setup environment
 * "print queues" can now display more information about queues, like free job slots, total running jobs